+++ src/traits/ActiveFormAjaxSubmitTrait.php
@@ -8,27 +8,72 @@
 
 namespace skeeks\cms\traits;
 
+use skeeks\cms\assets\ActiveFormAjaxSubmitAsset;
+
 /**
- * Class ActiveFormAjaxSubmit
+ * Trait ActiveFormAjaxSubmitTrait
  * @package skeeks\cms\traits
  */
 trait ActiveFormAjaxSubmitTrait
 {
+    /**
+     * @var null
+     */
+    public $clientCallback = null;
+
+    /**
+     * @deprecated
+     * @var string
+     */
+    public $afterValidateCallback = "";
+
     public function registerJs()
     {
+        ActiveFormAjaxSubmitAsset::register($this->view);
+
+        $this->view->registerJs(<<<JS
+sx.ActiveForm = new sx.classes.activeForm.AjaxSubmit('{$this->id}');
+JS
+    );
         $afterValidateCallback = $this->afterValidateCallback;
+        $clientCallback = $this->clientCallback;
+
+        if ($clientCallback) {
+            $this->view->registerJs(<<<JS
+            var callback = $clientCallback;
+            callback(sx.ActiveForm);
+JS
+    );
+        }
+        elseif ($afterValidateCallback) {
+            $this->view->registerJs(<<<JS
+            sx.ActiveForm.AjaxQueryHandler.set();
+            sx.ActiveForm.on('afterValidate', function() {
+                var callback = $afterValidateCallback;
+                callback(sx.ActiveForm.jForm, sx.ActiveForm.AjaxQuery);
+            });
+JS
+    );
+        }
+
+
+        /*$afterValidateCallback = $this->afterValidateCallback;
         if ($afterValidateCallback) {
             $this->view->registerJs(<<<JS
+            
                     
                 $('#{$this->id}').on('beforeSubmit', function (event, attribute, message) {
+                    console.log('beforeSubmit');
                     return false;
                 });
 
                 $('#{$this->id}').on('submit', function (event, attribute, message) {
+                    console.log('submit');
                     return false;
                 });
 
                 $('#{$this->id}').on('beforeValidate', function (event, messages, deferreds) {
+                    console.log('beforeValidate');
                 });
 
                 $('#{$this->id}').on('ajaxComplete', function (event, jqXHR, textStatus) {
@@ -44,6 +89,8 @@ trait ActiveFormAjaxSubmitTrait
 
                 $('#{$this->id}').on('afterValidate', function (event, messages, errorAttributes) {
 
+                    console.log('afterValidate');
+                    
                     if (_.size(errorAttributes) > 0)
                     {
                         sx.notify.error('Проверьте заполненные поля в форме');
@@ -105,7 +152,7 @@ JS
 
 JS
             );
-        }
+        }*/
 
     }